Add files via upload
The ui.common files are copies of the original .ui files. The .ui.macos files contain each one <false> to <true> change for getting the macOS native menu bar. The changes in cdbrowser fix the path for .css and .js files in the <app>.app/Contents/Resources folder under macOS.